+++ tvb_contrib/tvb/contrib/tests/cosimulation/simple/double_proxy_precision_complex_test.py
@@ -40,11 +40,11 @@ class TestDoubleProxyPrecisionComplex(BaseTestCase):
     def test_double_proxy_precision_complex(self):
         weight = np.array([[5, 2, 4, 0], [8, 5, 4, 1], [6, 1, 7, 9], [10, 0, 5, 6]])
         delay = np.array([[7, 8, 5, 1], [10, 3, 7, 9], [4, 3, 2, 8], [9, 10, 11, 5]])
-        max = np.int(np.max(delay)*10+1)
-        init_value = np.array([[[0.9,0.0], [0.1,0.0], [0.2,0.0], [0.3,0.0]]] * max)
+        max = np.int_(np.max(delay) * 10 + 1)
+        init_value = np.array([[[0.9, 0.0], [0.1, 0.0], [0.2, 0.0], [0.3, 0.0]]] * max)
         initial_condition = init_value.reshape((max, 2, weight.shape[0], 1))
         resolution_simulation = 0.1
-        synchronization_time =np.min(delay)
+        synchronization_time = np.min(delay)
         proxy_id_1 = [1]
         proxy_id_2 = [0, 2]
 
@@ -67,15 +67,15 @@ class TestDoubleProxyPrecisionComplex(BaseTestCase):
         time_ref, s_ref, result_ref = sim_ref(synchronization_time, rate=True)
 
         # COMPARE PROXY 1
-        np.testing.assert_array_equal(np.squeeze(result_ref[:,proxy_id_2,:], axis=2),
-                          np.squeeze(result_1[0][:,proxy_id_2,:], axis=2))
-        np.testing.assert_array_equal(np.squeeze(s_ref[:,proxy_id_2,:], axis=2),
-                            np.squeeze(s_1[0][:,proxy_id_2,:], axis=2))
+        np.testing.assert_array_equal(np.squeeze(result_ref[:, proxy_id_2, :], axis=2),
+                                      np.squeeze(result_1[0][:, proxy_id_2, :], axis=2))
+        np.testing.assert_array_equal(np.squeeze(s_ref[:, proxy_id_2, :], axis=2),
+                                      np.squeeze(s_1[0][:, proxy_id_2, :], axis=2))
         # COMPARE PROXY 2
-        np.testing.assert_array_equal(np.squeeze(result_ref[:,proxy_id_1,:], axis=2),
-                          np.squeeze(result_2[0][:,proxy_id_1,:], axis=2))
-        np.testing.assert_array_equal(np.squeeze(s_ref[:,proxy_id_1,:], axis=2),
-                            np.squeeze(s_2[0][:,proxy_id_1,:], axis=2))
+        np.testing.assert_array_equal(np.squeeze(result_ref[:, proxy_id_1, :], axis=2),
+                                      np.squeeze(result_2[0][:, proxy_id_1, :], axis=2))
+        np.testing.assert_array_equal(np.squeeze(s_ref[:, proxy_id_1, :], axis=2),
+                                      np.squeeze(s_2[0][:, proxy_id_1, :], axis=2))
 
         for i in range(0, 1000):
             time, s_2, result_2 = sim_2(synchronization_time,
@@ -83,7 +83,7 @@ class TestDoubleProxyPrecisionComplex(BaseTestCase):
 
             # compare with Raw monitor delayed by synchronization_time
             np.testing.assert_array_equal(result_ref[:, proxy_id_1, :], result_2[1][:, proxy_id_1, :])
-            np.testing.assert_array_equal(result_ref[:, proxy_id_2, :]*np.NAN, result_2[1][:, proxy_id_2, :])
+            np.testing.assert_array_equal(result_ref[:, proxy_id_2, :] * np.NAN, result_2[1][:, proxy_id_2, :])
             np.testing.assert_array_equal(s_ref, s_2[1])
 
             time, s_1, result_1 = sim_1(synchronization_time,
@@ -91,14 +91,14 @@ class TestDoubleProxyPrecisionComplex(BaseTestCase):
 
             # compare with Raw monitor delayed by synchronization_time
             np.testing.assert_array_equal(result_ref[:, proxy_id_2, :], result_1[1][:, proxy_id_2, :])
-            np.testing.assert_array_equal(result_ref[:, proxy_id_1, :]*np.NAN, result_1[1][:, proxy_id_1, :])
+            np.testing.assert_array_equal(result_ref[:, proxy_id_1, :] * np.NAN, result_1[1][:, proxy_id_1, :])
             np.testing.assert_array_equal(s_ref, s_1[1])
 
             time_ref, s_ref, result_ref = sim_ref(synchronization_time, rate=True)
             # COMPARE PROXY 1
             np.testing.assert_array_equal(
                 np.squeeze(result_ref[:, proxy_id_2, :], axis=2),
-                np.squeeze(result_1[0][:, proxy_id_2, :],  axis=2))
+                np.squeeze(result_1[0][:, proxy_id_2, :], axis=2))
             np.testing.assert_array_equal(
                 np.squeeze(s_ref[:, proxy_id_2, :], axis=2),
                 np.squeeze(s_1[0][:, proxy_id_2, :], axis=2))
